Use a dedicated config file
Piggy-backing on `git config` seemed like a good idea at the time, but
we can get much more fine-grained configuration if we maintain our own
config file:

* The configuration can be versioned
* We can have per-directory configurations (if different dirs in
  the same repo use different clang-format binaries [perverse, but not
  unthinkable])

Use the current directory as starting point, and search upwards for a
`git-format.config` file. If none is found, default to
`~/git-format.config`, and if that's not found, default to a simple
hard-coded configuration.

`git-format.config` uses a simple INI file format with a single section
`[format]`.