Apply formatting to the right revision of a file
Prior to this patch, git-format had a misinformed assumption that all
files in a diff exist in the working directory.

This is of course not always true (for example: 'git show HEAD~1000'
might have modified lines in a file that no longer exists, and we still
want to be able to format that diff).

To fix this, read the blob SHA from 'index' lines in the diff, and read
that blob out of the git database. If the blob does not exist, we assume
it's a local file in the working directory that hasn't yet been
committed.

This should be enough to run git-format over any diff, whether a
historical diff, a single commit or stage, or the current working
directory.