Add -p/--strip to strip diff paths
By default, git prefixes paths in diffs with 'a/' and 'b/'. git-format
used to assume this was always the case.

However, there's a global setting in git, diff.noprefix, that disables
this behavior for all diffs. Furthermore, 'git diff' takes --src-prefix
and --dst-prefix switches to pick a custom prefix, e.g. a/b/c/d/.

In the face of this variability, do what 'patch' would do: add a -p
switch so git-format users can compensate for any strange diff prefix
format.

Also print a hint when formatting fails due to path-not-found.